requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
frozenlist==1.2.0
future==0.18.2
idna==3.3
line-bot-sdk==2.0.1
multidict==5.2.0
numpy==1.21.4
opencv-python==4.5.4.60
PyYAML==6.0
requests==2.26.0
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
